requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
fastapi-jwt-auth==0.5.0
passlib==1.7.4
bcrypt==3.2.0
cryptography==35.0.0
python-dotenv==0.21.0
SQLAlchemy==1.4.36
python-jose==3.3.0
psycopg2==2.9.5
email-validator==1.3.1